public14/sentiment-compare/services/consumer/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.4.1+cpu
torchvision==0.19.1+cpu
transformers==4.44.2
sentencepiece==0.2.0
protobuf==5.27.2
psycopg2-binary
kafka-python-ng